fix: appdetailmodelsheetを削除して、２種類のファイルに分割しました
