+++ lib/ui/screen/detail/detail_post_screen.dart
@@ -11,7 +11,8 @@ import 'package:food_gram_app/main.dart';
 import 'package:food_gram_app/ui/component/app_heart.dart';
 import 'package:food_gram_app/ui/component/app_loading.dart';
 import 'package:food_gram_app/ui/component/dialog/app_share_dialog.dart';
-import 'package:food_gram_app/ui/component/modal_sheet/app_detail_modal_sheet.dart';
+import 'package:food_gram_app/ui/component/modal_sheet/app_detail_my_info_modal_sheet.dart';
+import 'package:food_gram_app/ui/component/modal_sheet/app_detail_other_info_modal_sheet.dart';
 import 'package:food_gram_app/utils/provider/loading.dart';
 import 'package:gap/gap.dart';
 import 'package:gif/gif.dart';
